images/starter/requirements.txt
boto3
botocore
click
Flask
Flask-SQLAlchemy
itsdangerous
Jinja2
jmespath
MarkupSafe
Pillow
python-dateutil
s3transfer
six
SQLAlchemy
urllib3
Werkzeug
